feat: change supabase booking table by passing breakfast data when user add breakfast
