+++ src/services/apiBookings.js
@@ -54,3 +54,20 @@ export async function getBookingDetail(id) {
 
   return data
 }
+
+export async function updatedBooking(id, obj) {
+  console.log(obj)
+  const { data, error } = await supabase
+    .from("bookings")
+    .update(obj)
+    .eq("id", id)
+    .select()
+    .single();
+
+  if (error) {
+    console.error(error);
+    throw new Error("Booking could not be updated");
+  }
+
+  return data;
+}
